+++ Products/CMFPlone/interfaces/__init__.py
@@ -23,6 +23,7 @@ from controlpanel import ITinyMCESchema
 from controlpanel import ITinyMCEToolbarSchema
 from controlpanel import ITypesSchema
 from controlpanel import IUserGroupsSettingsSchema
+from controlpanel import ISocialMediaSchema
 from events import IConfigurationChangedEvent
 from events import IReorderedEvent
 from events import ISiteManagerCreatedEvent